Move categoricalColors and transformDataTo100 to chartHelpers.js
